+++ scripts/build-scriptlets.js
@@ -1,6 +1,9 @@
-import { rollupStandard } from './rollup-runners';
-import { scriptlets, scriptletsListConfig } from '../rollup.config';
+import { rolldownStandard } from './rollup-runners';
+import { scriptletsListConfig, scriptletsConfig, typesConfig } from '../rollup.config';
 
-export const buildScriptletsList = async () => rollupStandard(scriptletsListConfig);
+export const buildScriptletsList = async () => rolldownStandard(scriptletsListConfig);
 
-export const buildScriptlets = async () => rollupStandard(scriptlets);
+export const buildScriptlets = async () => {
+    await rolldownStandard(scriptletsConfig);
+    await rolldownStandard(typesConfig);
+};